extract components for form management
