+++ app/components/form/instance.hbs
@@ -1,0 +1,47 @@
+<div {{did-insert this.onInit}}>
+  {{#if this.initialized}}
+    <div class="au-c-main-container">
+      <div
+        class="au-c-main-container__content au-c-main-container__content--scroll"
+      >
+        <div class="au-o-layout">
+          <section class="au-o-region-large">
+            <RdfForm
+              @groupClass="au-o-grid__item"
+              @form={{this.formInfo.form}}
+              @show={{false}}
+              @graphs={{this.formInfo.graphs}}
+              @sourceNode={{this.formInfo.sourceNode}}
+              @formStore={{this.formInfo.formStore}}
+              @forceShowErrors={{true}}
+              @useNewListingLayout={{true}}
+              @level={{2}}
+              class="au-u-max-width-medium"
+            />
+            {{outlet}}
+
+          </section>
+          <AuButton
+            {{on "click" this.saveInstance}}
+            @disabled={{this.isSaving}}
+          >
+            Bewaar
+          </AuButton>
+          <section class="au-o-region-large">
+            <AuHeading
+              @level="2"
+              @skin="3"
+              class="au-u-margin-bottom"
+            >Content</AuHeading>
+            <p>
+              {{this.sourceTriples}}
+            </p>
+          </section>
+        </div>
+      </div>
+    </div>
+
+  {{else}}
+    <div>loading...</div>
+  {{/if}}
+</div>
