+++ app/controllers/mandatenbeheer/fracties-new/instances.js
@@ -6,27 +6,12 @@ export default class FormInstancesController extends Controller {
   @service router;
 
   @action
-  createNewInstance() {
+  onCreate() {
     this.router.transitionTo('mandatenbeheer.fracties-new.new');
   }
 
   @action
-  async removeInstance(instanceId) {
-    const result = await fetch(
-      `/form-content/${this.model.formId}/instances/${instanceId}`,
-      {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/vnd.api+json',
-        },
-      }
-    );
-
-    if (!result.ok) {
-      this.errorMessage =
-        'Er ging iets mis bij het verwijderen van het formulier. Probeer het later opnieuw.';
-      return;
-    }
-    this.send('reloadModel');
+  async onRemoveInstance() {
+    this.router.refresh();
   }
 }
