+++ app/controllers/mandatenbeheer/fracties-new/new.js
@@ -1,62 +1,28 @@
 import Controller from '@ember/controller';
-import { tracked } from '@glimmer/tracking';
-import { action } from '@ember/object';
-import { keepLatestTask } from 'ember-concurrency';
 import { inject as service } from '@ember/service';
+import { action } from '@ember/object';
 
 export default class FormNewController extends Controller {
-  @tracked sourceTriples;
-  @tracked errorMessage;
   @service router;
 
-  registerObserver() {
-    this.model.formStore.registerObserver(() => {
-      this.sourceTriples = this.model.formStore.serializeDataMergedGraph(
-        this.model.graphs.sourceGraph
-      );
-    });
-  }
+  // while not technically an 'action' we need this to bind the function to the controller's scope
+  @action
+  buildSourceTtl(instanceUri) {
+    const mandatenbeheer = this.model.mandatenbeheer;
+    const bestuurseenheid = mandatenbeheer.bestuurseenheid;
+    const bestuursorganen = mandatenbeheer.bestuursorganen;
 
-  get isSaving() {
-    return this.save.isRunning;
-  }
+    const bestuurseenheidUri = bestuurseenheid.get('uri');
+    const bestuursOrganenUris = bestuursorganen.map((b) => `<${b.get('uri')}>`);
 
-  @keepLatestTask
-  *save() {
-    const triples = this.sourceTriples;
-    const definition = this.model.definition;
-    this.errorMessage = null;
-    // post triples to backend
-    const result = yield fetch(`/form-content/${definition.id}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/vnd.api+json',
-      },
-      body: JSON.stringify({
-        contentTtl: triples,
-        instanceUri: this.model.sourceNode.value,
-      }),
-    });
-
-    if (!result.ok) {
-      this.errorMessage =
-        'Er ging iets mis bij het opslaan van het formulier. Probeer het later opnieuw.';
-      return;
-    }
-
-    const { id } = yield result.json();
-
-    if (!id) {
-      this.errorMessage =
-        'Het formulier werd niet correct opgeslagen. Probeer het later opnieuw.';
-      return;
-    }
-
-    this.router.transitionTo('form.instance', definition.id, id);
+    return `<${instanceUri}> <http://www.w3.org/ns/org#memberOf> ${bestuursOrganenUris.join(
+      ', '
+    )} .
+    <${instanceUri}> <http://www.w3.org/ns/org#linkedTo> <${bestuurseenheidUri}> . `;
   }
 
   @action
-  async createInstance() {
-    this.save.perform();
+  onCreate(id) {
+    this.router.transitionTo('mandatenbeheer.fracties-new.instance', id);
   }
 }
