+++ app/router.js
@@ -37,7 +37,7 @@ Router.map(function () {
     this.route('fracties-new', function () {
       this.route('new');
       this.route('instances');
-      this.route('instance', { path: '/fractie/:instance_id' });
+      this.route('instance', { path: '/:instance_id' });
       this.route('edit');
     });
   });
