+++ app/templates/mandatenbeheer/fracties-new/new.hbs
@@ -1,47 +1,6 @@
 {{page-title "New"}}
-<div class="au-c-main-container">
-  <div
-    class="au-c-main-container__content au-c-main-container__content--scroll"
-  >
-    <div class="au-o-layout">
-      <section class="au-o-region-large">
-        <RdfForm
-          @groupClass="au-o-grid__item"
-          @form={{@model.form}}
-          @show={{false}}
-          @graphs={{this.model.graphs}}
-          @sourceNode={{this.model.sourceNode}}
-          @formStore={{this.model.formStore}}
-          @forceShowErrors={{true}}
-          @useNewListingLayout={{true}}
-          @level={{2}}
-          class="au-u-max-width-medium"
-        />
-        {{outlet}}
-
-      </section>
-      {{#if this.errorMessage}}
-        <AuAlert @skin="error" @title="Fout" @icon="alert-triangle">
-          {{this.errorMessage}}
-        </AuAlert>
-      {{/if}}
-
-      <AuButton {{on "click" this.createInstance}} @disabled={{this.isSaving}}>
-        Bewaar
-      </AuButton>
-
-      <section class="au-o-region-large">
-
-        <AuHeading
-          @level="2"
-          @skin="3"
-          class="au-u-margin-bottom"
-        >Content</AuHeading>
-        <p>
-          {{this.sourceTriples}}
-        </p>
-
-      </section>
-    </div>
-  </div>
-</div>
+<Form::NewInstance
+  @onCreate={{this.onCreate}}
+  @form={{this.model.form}}
+  @buildSourceTtl={{this.buildSourceTtl}}
+/>
